Create week3_points
Question: Points
Define class Point, representing a point on the 2-dimensional plane. The class should contain two fields, representing the coordinates of these points: x and y.

Define a function that takes two Point objects (with x, y coordinates) and returns whichever one is closest to the origin. Do so by defining a method in the Point class that computes the distance to the origin.

Input
The first line of the input contains T: the number of test cases. T test cases follow.

Each test case is a line containing four integer numbers: the x coorinate of the first point, followed by the y coordinate of the first point, followed by the x coorinate of the second point, followed by the y coordinate of the second point.

Output
For each test case, output one line containing Case #t: x y, where t is the test case number (starting from 1), and x and y are the coordinates of the point closes to the origin.