Create week3_reverse
Problem: Reverse
Write a function that takes a string and reverses every word in it. Reversed word is a word which is written from back to front.

Input
The first line of the input contains T: the number of test cases. T test cases follow.

Each test case is a line containing one or more words. Every word is a sequence of small latin characters. The words are separated by a single space character.

Output
For each test case, output one line containing Case #x: S, where x is the test case number (starting from 1), and S is the corresponding string with all words reversed.