Create week4_reverse
Problem: Extremes
Given an array, count how many elements in the array are equal to the maximal or to the minimal value of the array.

Input
The first line of the input contains T: the number of test cases. T test cases follow.

Each test case is a line containing N: the number of elements in the array, followed by N integers: the elements of the array.

Output
For each test case, output one line containing Case #x: y, where x is the test case number (starting from 1), and y is the number of elements in the array which are equal to the maximal or to the minimal value of the array.